apply file separations, no default exports
