+++ src/filters/filter.ts
@@ -1,0 +1,6 @@
+import { CommandHandler } from "../commandsHandlers/commandHandler";
+
+export interface Filter {
+	apply(handler: CommandHandler): CommandHandler
+}
+
